+++ src/java/com/sapienter/jbilling/server/process/task/PaymentTermPenaltyTask.java
@@ -20,24 +20,23 @@
 
 package com.sapienter.jbilling.server.process.task;
 
-import java.math.BigDecimal;
-
-import org.apache.log4j.Logger;
-
 import com.sapienter.jbilling.server.invoice.NewInvoiceDTO;
 import com.sapienter.jbilling.server.order.TimePeriod;
 import com.sapienter.jbilling.server.pluggableTask.TaskException;
 import com.sapienter.jbilling.server.pluggableTask.admin.ParameterDescription;
 import com.sapienter.jbilling.server.util.Constants;
 import com.sapienter.jbilling.server.util.MapPeriodToCalendar;
+import org.apache.log4j.Logger;
+
+import java.math.BigDecimal;
 
 /**
  * This plug-in calculates taxes for invoice.
  *
  * Plug-in parameters:
- * 
+ *
  *      penalty_after_days: (required) Number of days beyond which the penalty becomes applicable
- * 
+ *
  * @author Vikas Bodani
  * @since 28-Jul-2011
  *
@@ -45,32 +44,32 @@ import com.sapienter.jbilling.server.util.MapPeriodToCalendar;
 public class PaymentTermPenaltyTask extends AbstractChargeTask {
 
     private static final Logger LOG = Logger.getLogger(PaymentTermPenaltyTask.class);
-    
+
     // Plug-in Parameters
     // Mandatory parameters
     protected static final ParameterDescription PARAM_PENALTY_AFTER_DAYS = new ParameterDescription("penalty_after_days", true, ParameterDescription.Type.STR);
-    
-    private Integer penaltyAfterDays=null; 
-    
+
+    private Integer penaltyAfterDays=null;
+
     //initializer for pluggable params
     {
         descriptions.add(PARAM_PENALTY_AFTER_DAYS);
     }
-    
+
     /**
-     * 
+     *
      */
-    protected BigDecimal calculateAndApplyTax(NewInvoiceDTO invoice, Integer userId) { 
-        
+    protected BigDecimal calculateAndApplyTax(NewInvoiceDTO invoice, Integer userId) {
+
         LOG.debug("calculateAndApplyTax");
-        
+
         BigDecimal invoiceAmountSum= super.calculateAndApplyTax(invoice, userId);
-        
+
         this.invoiceLineTypeId= Constants.INVOICE_LINE_TYPE_PENALTY;
-        
+
         return invoiceAmountSum;
     }
-    
+
     /**
      * Custom logic to determine if the tax should be applied to this user's invoice
      * @param userId The user_id of the Invoice
@@ -84,13 +83,13 @@ public class PaymentTermPenaltyTask extends AbstractChargeTask {
         //convert period unit to days (Days, Weeks, Months, Years to days)
         int periodToDays= MapPeriodToCalendar.periodToDays(timePeriod.getUnitId());
         //Period unit value
-        int periodValue= timePeriod.getValue().intValue();
+        int periodValue= timePeriod.getValue();
         //product of the above two to get total number of days
-        Integer totalDays= Integer.valueOf(periodValue*periodToDays);
+        Integer totalDays= periodValue * periodToDays;
         LOG.debug("Total Invoice Due days= " + totalDays);
         return (totalDays.compareTo(penaltyAfterDays) > 0 );
     }
-    
+
     /**
      * Set the current set of plugin params
      */
@@ -110,5 +109,5 @@ public class PaymentTermPenaltyTask extends AbstractChargeTask {
             throw new TaskException(e);
         }
     }
-    
+
 }
